seiralise NanoCPUs as string
